+++ spec/support/output_writer.rb
@@ -1,11 +1,8 @@
 shared_examples_for "output writer" do
   before do
     @read, @write = IO.pipe
-  end
-
-  let(:browser) do
-    connection = Capybara::Webkit::Connection.new(:stderr => @write)
-    Capybara::Webkit::Browser.new(connection)
+    configure { |config| config.stderr = @write }
+    fork_connection
   end
 
   let(:stderr) do